+++ src/SearchSystem.VectorSearch/TermEntryStats.cs
@@ -1,5 +1,8 @@
 using System;
+using System.Text.RegularExpressions;
 using SearchSystem.Infrastructure.Documents;
+using SearchSystem.Infrastructure.Extensions;
+using Sprache;
 
 // ReSharper disable BuiltInTypeReferenceStyle
 using Term = System.String;
@@ -69,5 +72,43 @@ namespace SearchSystem.VectorSearch
 		/// <inheritdoc />
 		public override string ToString()
 			=> $"{Term,32} {DocumentLink.Name,8} {TermFrequency,12:F8} {InverseDocFrequency,12:F8} {TfIdf,12:F8}";
+
+		/// <summary>
+		/// Parse <paramref name="input"/> to <see cref="TermEntryStats"/> instance. 
+		/// </summary>
+		public static TermEntryStats ParseLine(string input)
+			=> input
+				.Trim()
+				.To(str => Regex.Replace(str, @"\s+", " "))
+				.To(TermStatsGrammar.SingleEntry.Parse);
+
+		/// <summary>
+		/// Definition of serialized stats grammar.
+		/// </summary>
+		private static class TermStatsGrammar
+		{
+			/// <remarks>
+			/// Stats is always calculated based on normalized documents.
+			/// </remarks>
+			private const string subsectionName = "Normalization";
+
+			/// <summary>
+			/// Single <see cref="TermEntryStats"/> parser.
+			/// </summary>
+			public static Parser<TermEntryStats> SingleEntry =>
+				from term                in Parse.CharExcept(' ').Many().Text()
+				from documentLink        in Parse
+					.Regex(@"^d+\.[a-z]{3}$")
+					.Token()
+					.Select(fileName => new DocumentLink(subsectionName, fileName))
+				from termFrequency       in DoubleParser
+				from inverseDocFrequency in DoubleParser
+				select new TermEntryStats(term, documentLink, termFrequency, inverseDocFrequency);
+
+			/// <summary>
+			/// Double value parser.
+			/// </summary>
+			private static Parser<double> DoubleParser => Parse.DecimalInvariant.Select(double.Parse).Token();
+		}
 	}
 }
